Rename GitHubClient as VCS::GitHub
Since I sometimes confused by the difference between Client and
GitHubClient.
