--- lib/circleci/coverage_reporter/abstract_vcs_client.rb
@@ -1,21 +1,0 @@
-module CircleCI
-  module CoverageReporter
-    # @abstract Subclass and override {#create_comment} to implement a custom VCS client class.
-    class AbstractVCSClient
-      # @param token [String]
-      def initialize(token)
-        @token = token
-      end
-
-      # @param body [String]
-      # @return [void]
-      def create_comment(body) # rubocop:disable Lint/UnusedMethodArgument
-        raise NotImplementedError
-      end
-
-      private
-
-      attr_reader :token
-    end
-  end
-end
